% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_byrow.R
\name{align_byrow}
\alias{align_byrow}
\title{Align objects by their row names}
\usage{
align_byrow(..., return_index = FALSE)
}
\arguments{
\item{...}{A set of objects: vectors, lists, matrices, data frames, and/or arrays}

\item{return_index}{If true, just return a vector with the
individuals that are in common across all objects.}
}
\value{
If \code{return_index=FALSE} (the default), a list
containing the objects reduced to the individuals in common across
all of them. If \code{return_index=TRUE}, just return a vector of
character strings for the individuals that are in common.
}
\description{
Align a set of objects by their row names, omitting rows that are
not present in all objects.
}

