% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.cross2.R
\name{basic_summaries}
\alias{basic_summaries}
\alias{chr_names}
\alias{covar_names}
\alias{ind_ids}
\alias{line_ids}
\alias{marker_names}
\alias{n_chr}
\alias{n_covar}
\alias{n_ind}
\alias{n_lines}
\alias{n_mar}
\alias{n_pheno}
\alias{n_phenocovar}
\alias{pheno_names}
\alias{phenocovar_names}
\alias{tot_mar}
\title{Basic summaries of a cross2 object}
\usage{
n_ind(cross2)

ind_ids(cross2)

n_lines(cross2)

line_ids(cross2)

n_chr(cross2)

chr_names(cross2)

tot_mar(cross2)

n_mar(cross2)

marker_names(cross2)

n_pheno(cross2)

pheno_names(cross2)

n_covar(cross2)

covar_names(cross2)

n_phenocovar(cross2)

phenocovar_names(cross2)
}
\arguments{
\item{cross2}{An object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}
}
\description{
Basic summaries of a cross2 object.
}
\section{Functions}{
\itemize{
\item \code{n_ind}: Number of phenotyped individuals

\item \code{ind_ids}: IDs of phenotyped individuals

\item \code{n_lines}: Number of genotyped lines

\item \code{line_ids}: IDs of genotyped lines

\item \code{n_chr}: Number of chromosomes

\item \code{chr_names}: Chromosome names

\item \code{tot_mar}: Total number of markers

\item \code{n_mar}: Number of markers on each chromosome

\item \code{marker_names}: Marker names

\item \code{n_pheno}: Number of phenotypes

\item \code{pheno_names}: Phenotype names

\item \code{n_covar}: Number of covariates

\item \code{covar_names}: Covariate names

\item \code{n_phenocovar}: Number of phenotype covariates

\item \code{phenocovar_names}: Phenotype covariate names
}}
\seealso{
\code{\link{summary.cross2}}
}
\keyword{utilities}

