% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kinship.R
\name{calc_kinship}
\alias{calc_kinship}
\title{Calculate kinship matrix}
\usage{
calc_kinship(probs, type = c("overall", "loco", "chr"),
  use_grid_only = TRUE, omit_x = FALSE, use_allele_probs = TRUE,
  normalize = TRUE, quiet = TRUE, cores = 1)
}
\arguments{
\item{probs}{List of three-dimensional arrays of probabilities, as
calculated from \code{\link{calc_genoprob}}.}

\item{type}{Indicates whether to calculate the overall kinship
(\code{"overall"}, using all chromosomes), the kinship matrix
leaving out one chromosome at a time (\code{"loco"}), or the
kinship matrix for each chromosome (\code{"chr"}).}

\item{use_grid_only}{If \code{TRUE} and \code{probs} were calculated with
\code{stepwidth="fixed"}, reduce them to the grid using
\code{\link{probs_to_grid}}.}

\item{omit_x}{If \code{TRUE}, only use the autosomes; ignored when
\code{type="chr"}.}

\item{use_allele_probs}{If \code{TRUE}, assess similarity with
allele probabilities (that is, first run
\code{\link{genoprob_to_alleleprob}}); otherwise use the genotype
probabilities.}

\item{normalize}{If \code{TRUE}, divide the kinship matrix by a
normalizing constant (see Details), to give better estimates of
heritability.}

\item{quiet}{IF \code{FALSE}, print progress messages.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}
}
\value{
If \code{type="overall"} (the default), a matrix of
proportion of matching alleles. Otherwise a list with one matrix
per chromosome.
}
\description{
Calculate genetic similarity among individuals (kinship matrix)
from conditional genotype probabilities.
}
\details{
If \code{use_allele_probs=TRUE} (the default), we first
convert the genotype probabilities are converted to allele
probabilities (using \code{\link{genoprob_to_alleleprob}}). This is
recommended, as then the result is twice the empirical kinship
coefficient (e.g., the expected value for an intercross is 1/2;
using genotype probabilities, the expected value is 3/8).

We then calculate
\eqn{\sum_{kl}(p_{ikl} p_{jkl})}{sum_kl (p_ikl p_jkl)}
where \eqn{k} = position, \eqn{l} = allele, and \eqn{i,j}
are two individuals.

For crosses with just two possible genotypes (e.g., backcross), we
don't convert to allele probabilities but just use the original
genotype probabilities.

If \code{normalize=TRUE} (the default), we normalize the kinship matrix as in
equation 5 in Kang et al. (2010) Nat Genet
42:348-354. \href{http://www.nature.com/ng/journal/v42/n4/full/ng.548.html}{doi: 10.1038/ng.548}
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
probs <- calc_genoprob(grav2, step=1, error_prob=0.002)
sim <- calc_kinship(probs)
}
\keyword{utilities}

