% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_raw_summaries.R
\name{calc_raw_het}
\alias{calc_raw_het}
\title{Calculate estimated heterozygosity from raw SNP genotypes}
\usage{
calc_raw_het(cross)
}
\arguments{
\item{cross}{Object of class \code{"cross2"}. For details, see the
\href{https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}
}
\value{
A vector of heterozygosities, one for each individual.
}
\description{
Calculate estimated heterozygosity for each individual from raw SNP genotypes
}
\examples{
\dontrun{
# load example data and calculate genotype probabilities
file <- paste0("https://raw.githubusercontent.com/rqtl/",
               "qtl2data/master/DOex/DOex.zip")
DOex <- read_cross2(file)
DOex_het <- calc_raw_het(DOex)
}
}
\keyword{utilities}
