% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_phasedgeno.R
\name{cbind.phasedgeno}
\alias{cbind.phasedgeno}
\title{Join phased genotype results for different chromosomes}
\usage{
\method{cbind}{phasedgeno}(...)
}
\arguments{
\item{...}{Imputed genotype objects as produced by
\code{\link[=guess_phase]{guess_phase()}}. Must have the same set of individuals.}
}
\value{
An object of class \code{"phasedgeno"}, like the input; see \code{\link[=guess_phase]{guess_phase()}}.
}
\description{
Join multiple phased genotype objects, as produced by \code{\link[=guess_phase]{guess_phase()}}, for the
same set of individuals but different chromosomes.
}
\examples{
\dontrun{
file <- paste0("https://raw.githubusercontent.com/rqtl/",
               "qtl2data/main/DOex/DOex.zip")
DOex <- read_cross2(file)
pr <- calc_genoprob(DOex, error_prob=0.002)
m <- maxmarg(pr)
phA <- guess_phase(DOex[,2], m[,2])
phB <- guess_phase(DOex[,3], m[,3])
ph <- cbind(phA, phB)
}

}
\seealso{
\code{\link[=rbind.phasedgeno]{rbind.phasedgeno()}}, \code{\link[=subset.phasedgeno]{subset.phasedgeno()}}, \code{\link[=guess_phase]{guess_phase()}}
}
