% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbind_scan1perm.R
\name{cbind.scan1perm}
\alias{cbind.scan1perm}
\title{Combine columns from multiple scan1 permutation results}
\usage{
\method{cbind}{scan1perm}(...)
}
\arguments{
\item{...}{A set of permutation results from
\code{\link{scan1perm}} (objects of class \code{scan1perm}). If
different numbers of permutation replicates were used, those
columns with fewer replicates are padded with missing values
\code{NA}. However, if any include autosome/X
chromosome-specific permutations, they must all be such.}
}
\value{
The combined column-binded input, as a \code{scan1perm} object.
}
\description{
Column-bind multiple scan1perm objects with the same numbers of rows.
}
\details{
The aim of this function is to concatenate the results
from multiple runs of a permutation test with
\code{\link{scan1perm}}, generally with different phenotypes
and/or methods, to be used in parallel with
\code{\link{rbind.scan1perm}}.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# insert pseudomarkers into map
map <- insert_pseudomarkers(iron$gmap, step=1)

# calculate genotype probabilities
probs <- calc_genoprob(iron, map, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# strata for permutations
perm_strata <- mat2strata(Xcovar)

# permutations with genome scan
\dontrun{
operm1 <- scan1perm(probs, pheno[,1,drop=FALSE], addcovar=covar, Xcovar=Xcovar,
                n_perm=1000, perm_Xsp=TRUE, perm_strata=perm_strata,
                chr_lengths=chr_lengths(iron$gmap))
operm2 <- scan1perm(probs, pheno[,2,drop=FALSE], addcovar=covar, Xcovar=Xcovar,
                n_perm=1000, perm_Xsp=TRUE, perm_strata=perm_strata,
                chr_lengths=chr_lengths(iron$gmap))
}

\dontshow{
operm1 <- scan1perm(probs, pheno[,1,drop=FALSE], addcovar=covar, Xcovar=Xcovar, n_perm=3, perm_strata=perm_strata)
operm2 <- scan1perm(probs, pheno[,2,drop=FALSE], addcovar=covar, Xcovar=Xcovar, n_perm=3, perm_strata=perm_strata)
}

operm <- cbind(operm1, operm2)

}
\seealso{
\code{\link{rbind.scan1perm}}, \code{\link{scan1perm}}, \code{\link{scan1}}
}
