% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_viterbi.R
\name{cbind.viterbi}
\alias{cbind.viterbi}
\title{Join viterbi results for different chromosomes}
\usage{
cbind.viterbi(...)
}
\arguments{
\item{...}{Imputed genotype objects as produced by
\code{\link{viterbi}}. Must have the same set of individuals.}
}
\value{
A single imputed genotype object.
}
\description{
Join multiple viterbi objects, as produced by
\code{\link{viterbi}} for different individuals.
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
gA <- viterbi(grav2[1:5,1:2], step=1, error_prob=0.002)
gB <- viterbi(grav2[1:5,3:4], step=1, error_prob=0.002)
g <- cbind(gA, gB)

}

