% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean an object}
\usage{
clean(object, ...)
}
\arguments{
\item{object}{Object to be cleaned}

\item{...}{Other arguments}
}
\value{
Input object with messy values cleaned up
}
\description{
Clean an object by removing messy values
}
\seealso{
\code{\link[=clean.scan1]{clean.scan1()}}, \code{\link[=clean.calc_genoprob]{clean.calc_genoprob()}}
}
