% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_genoprob.R
\name{compare_genoprob}
\alias{compare_genoprob}
\title{Compare two sets of genotype probabilities}
\usage{
compare_genoprob(probs1, probs2, cross, ind = 1, chr = NULL,
  minprob = 0.95, minmarkers = 10)
}
\arguments{
\item{probs1}{Genotype probabilities (as produced by \code{\link[qtl2geno:calc_genoprob]{qtl2geno::calc_genoprob()}})
or allele dosages (as produced by \code{\link[qtl2geno:genoprob_to_alleleprob]{qtl2geno::genoprob_to_alleleprob()}}).}

\item{probs2}{A second set of genotype probabilities, just like \code{probs1}.}

\item{cross}{Object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}

\item{ind}{Individual to plot, either a numeric index or an ID.}

\item{chr}{Selected chromosome; a single character string.}

\item{minprob}{Minimum probability for inferring genotypes (passed to \code{\link[=maxmarg]{maxmarg()}}).}

\item{minmarkers}{Minimum number of markers in results.}

\item{map_type}{Whether to use the physical or genetic map for
widths and positions (if selected map is not available, use
whichever one is available.}

\item{minwidth}{Minimum width in results.}
}
\value{
An object of class \code{"compare_genoprob}", which is a data frame with
each row corresponding to an interval over which \code{probs1} and
\code{probs2} each have a fixed inferred genotype. Columns include the
two inferred genotypes, the start and end points and width of the
interval, and when founder genotypes are in \code{cross}, the
proportions of SNPs where the individual matches each possible
genotypes.
}
\description{
Compare two sets of genotype probabilities for one individual on a single chromosome.
}
\details{
The function does the following:
\itemize{
\item Reduce the probabilities to a set of common locations that also appear in \code{cross}.
\item Use \code{\link[=maxmarg]{maxmarg()}} to infer the genotype at every position using each set of probabilities.
\item Identify intervals where the two inferred genotypes are constant.
\item Within each segment, compare the observed SNP genotypes to the founders' genotypes.
}
}
\seealso{
\code{\link[qtl2plot:plot_genoprobcomp]{qtl2plot::plot_genoprobcomp()}}
}
