% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2scanone.R
\name{convert2scanone}
\alias{convert2scanone}
\title{Convert scan1 results to the scanone format}
\usage{
convert2scanone(output)
}
\arguments{
\item{output}{Matrix of LOD scores, as calculated by
\code{\link{scan1}} or \code{\link{scan1_lmm}}.}
}
\value{
A data frame with class \code{"scanone"}, containing
chromosome and position columns followed by the LOD scores in
\code{output}.
}
\description{
Convert the results of \code{\link{scan1}} or
\code{\link{scan1_lmm}} to the form used by the R/qtl function
\code{scanone}.
}
\examples{
library(qtl2geno)
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
probs <- calc_genoprob(iron, step=1, error_prob=0.002)
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)
out <- scan1(probs, pheno, covar, Xcovar)

out_rev <- convert2scanone(out)

}

