% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_xo.R
\name{count_xo}
\alias{count_xo}
\title{Count numbers of crossovers}
\usage{
count_xo(geno, quiet = TRUE, cores = 1)
}
\arguments{
\item{geno}{List of matrices of genotypes (output of \code{\link{maxmarg}} or \code{\link{viterbi}}).}

\item{quiet}{If FALSE, print progress messages.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}
}
\value{
A matrix of crossover counts, individuals x chromosomes
}
\description{
Estimate the numbers of crossovers in each individual on each chromosome.
}
\examples{
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
map <- insert_pseudomarkers(iron$gmap, step=1)
pr <- calc_genoprob(iron, map, error_prob=0.002, map_function="c-f")
g <- maxmarg(pr)
count_xo(g)

}
