% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_query_func.R
\name{create_gene_query_func}
\alias{create_gene_query_func}
\title{Create a function to query genes}
\usage{
create_gene_query_func(dbfile = NULL, db = NULL, table_name = "genes",
  chr_field = "chr", start_field = "start_Mbp", stop_field = "stop_Mbp",
  filter = NULL)
}
\arguments{
\item{dbfile}{Name of database file}

\item{db}{Optional database connection (provide one of \code{file} and \code{db}).}

\item{table_name}{Name of table in the database}

\item{chr_field}{Name of chromosome field}

\item{start_field}{Name of field with start position (in Mbp)}

\item{stop_field}{Name of field with stop position (in Mbp)}

\item{filter}{Additional SQL filter (as a character string).}
}
\value{
Function with three arguments, \code{chr}, \code{start},
    and \code{end}, which returns a data frame with the genes
    overlapping that region. The output should contain at least the
    columns \code{Name}, \code{chr}, \code{start}, and \code{stop}, the latter
    two being positions in Mbp.
}
\description{
Create a function that will connect to a SQLite database of gene
information and return a data frame with gene information for a
selected region.
}
\examples{
\dontrun{
# create query function by connecting to file
dbfile <- system.file("extdata", "mouse_genes_small.sqlite", package="qtl2db")
query_genes <- create_gene_query_func(dbfile, filter="(source=='MGI')")
# query_genes will connect and disconnect each time
genes <- query_genes("2", 97.0, 98.0)

# connect and disconnect separately
library(RSQLite)
db <- dbConnect(SQLite(), dbfile)
query_genes <- create_gene_query_func(db=db, filter="(source=='MGI')")
genes <- query_genes("2", 97.0, 98.0)
dbDisconnect(db)
}
}
