% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_query_func.R
\name{create_snp_query_func}
\alias{create_snp_query_func}
\title{Create a function to query SNPs}
\usage{
create_snp_query_func(dbfile = NULL, db = NULL, table = "snps",
  chr_field = "chr", pos_field = "pos_Mbp")
}
\arguments{
\item{dbfile}{Name of database file}

\item{db}{Optional database connection (provide one of \code{file} and \code{db}).}

\item{table}{Name of table in the database}

\item{chr_field}{Name of chromosome field}

\item{pos_field}{Name of position field}
}
\value{
Function with three arguments, \code{chr}, \code{start}, and \code{end},
which returns a data frame with the SNPs in that region. The output should contain
at least the columns \code{chr} and \code{pos}, the latter being position in Mbp.
}
\description{
Create a function that will connect to a SQLite database of founder SNP
information and return a data frame with SNP information for a
selected region.
}
\examples{
\dontrun{
# create query function by connecting to file
query_snps <- create_snp_query_func("ccvariants.sqlite")
# query_snps will connect and disconnect each time
snps <- query_snps("19", 25.1, 26.1)

# connect and disconnect separately
db <- dbConnect(SQLite(), "ccvariants.sqlite")
query_snps <- create_snp_query_func(db=db)
snps <- query_snps("19", 25.1, 26.1)
dbDisconnect(db)
}
}
