% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est_map}
\alias{est_map}
\title{Estimate genetic maps}
\usage{
est_map(cross, error_prob = 0.0001, map_function = c("haldane", "kosambi",
  "c-f", "morgan"), maxit = 10000, tol = 0.000001, quiet = TRUE)
}
\arguments{
\item{cross}{Object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}

\item{error_prob}{Assumed genotyping error probability}

\item{map_function}{Character string indicating the map function to
use to convert genetic distances to recombination fractions.}

\item{maxit}{Maximum number of iterations in EM algorithm.}

\item{tol}{Tolerance for determining convergence}

\item{quiet}{If true, don't print any messages}
}
\description{
Uses a hidden Markov model to re-estimate the genetic map for an
experimental cross, with possible allowance for genotyping errors.
}

