% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoprob_to_snpprob.R
\name{genoprob_to_snpprob}
\alias{genoprob_to_snpprob}
\title{Convert genotype probabilities to SNP probabilities}
\usage{
genoprob_to_snpprob(genoprobs, snpinfo, tol = 0.00000001)
}
\arguments{
\item{genoprobs}{List of 3d arrays of genotype probabilities, as
calculated by \code{\link[qtl2geno]{calc_genoprob}}; each component
is a chromosome. Requires an attributes \code{"map"} with map
positions, \code{"alleles"} with the set of alleles, and
\code{"is_x_chr"} a logical vector indicating which chromosomes are
the X chromosome.}

\item{snpinfo}{Data frame with SNP information with the following columns:
\itemize{
\item \code{chr} - Character string or factor with chromosome
\item \code{pos} - Position (in same units as in the \code{"map"}
    attribute in \code{genoprobs}.
\item \code{sdp} - Strain distribution pattern: an integer, between
    1 and \eqn{2^n - 2} where \eqn{n} is the number of strains, whose
    binary encoding indicates the founder genotypes
\item \code{snp} - Character string with SNP identifier (if
    missing, the rownames are used).
}}

\item{tol}{Tolerance for determining whether a SNP is exactly at a
position at which genotype probabilities were already calculated.}
}
\value{
A List of 3d arrays of SNP genotype probabilities, similar
to the input. Each component (corresponding to a chromosome) has an
attribute \code{"snpinfo"} containing the input \code{snpinfo} with
an additional column, \code{index}, that is an index to the
genotype probabilities. (In cases where multiple SNPs are contained
in the same interval and have the same strain distribution pattern
(SDP), the probabilities would be all the same and so the results
contain the probabilities for just one of them.)

If the input \code{genoprobs} is for allele probabilities, the
output has just two probability columns (for the two SNP
alleles). If the input has a full set of \eqn{n(n+1)/2}
probabilities for \eqn{n} strains, the output has 3 probabilities
(for the three SNP genotypes). If the input has full genotype
probabilities for the X chromosome (\eqn{n(n+1)/2} genotypes for
the females followed by \eqn{n} hemizygous genotypes for the
males), the output has 5 probabilities: the 3 female SNP genotypes
followed by the two male hemizygous SNP genotypes.
}
\description{
For multi-parent populations, convert use founder genotypes at a
set of SNPs to convert founder-based genotype probabilities to SNP
genotype probabilities.
}
\details{
We first split the SNPs by chromosome and identify the
intervals in the genotype probabilities that contain each. For SNPs
within \code{tol} of a position at which the genotype probabilities
were calculated, we use those probabilities. For SNPs contained
within an interval, we use the average of the probabilities for the
two endpoints. We then collapse the probabilities according to the
strain distribution pattern.
}
\examples{
library(qtl2geno)
# show something here

}

