% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_pseudomarkers}
\alias{insert_pseudomarkers}
\title{Insert pseudomarkers into a marker map}
\usage{
insert_pseudomarkers(map, step = 0, off_end = 0, stepwidth = c("fixed",
  "max"), tol = 0.01, pmar_stem = "loc")
}
\arguments{
\item{map}{Numeric vector of marker positions, for a single
chromosome.}

\item{step}{Distance between pseudomarkers and markers; if
\code{step=0} no pseudomarkers are inserted.}

\item{off_end}{Distance beyond terminal markers in which to insert
pseudomarkers.}

\item{stepwidth}{Indicates whether to use a fixed grid
(\code{stepwidth="fixed"}) or to use the maximal distance between
pseudomarkers to ensure that no two adjacent markers/pseudomarkers
are more than \code{step} apart.}

\item{tol}{Tolerance for determining whether a pseudomarker would duplicate a marker position.}

\item{pmar_stem}{Character string to serve as the stem for naming the pseudomarkers.}
}
\value{
A vector of positions of pseudomarkers and markers: the
input \code{map} vector with pseudomarker positions added. An
attribute \code{"index"} is an integer vector that indicates which
positions are pseudomarkers (value 0) and which are markers
(positive values, indicating the marker indices). If
\code{stepwidth=fixed}, a further attributed (\code{"grid"}) is a
logical vector that indicates which positions correspond to the
fixed grid.
}
\description{
Insert pseudomarkers into a map of genetic markers, for a single chromosome.
}
\details{
If \code{stepwidth="fixed"}, a grid of pseudomarkers is added to the marker map.

If \code{stepwidth="max"}, a minimal set of pseudomarkers are
added, so that the maximum distance between adjacent markers or
pseudomarkers is at least \code{step}. If two adjacent markers are
separated by less than \code{step}, no pseudomarkers will be added
to the interval. If they are more then \code{step} apart, a set of
equally-spaced pseudomarkers will be added.
}
\examples{
library(qtl)
data(hyper)
chr4map <- pull.map(hyper, chr=4)
pmap <- insert_pseudomarkers(chr4map, step=1, pmar_stem="c4.loc")
}
\keyword{utilities}

