% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_map.R
\name{interp_map}
\alias{interp_map}
\title{Interpolate between maps}
\usage{
interp_map(map, oldmap, newmap)
}
\arguments{
\item{map}{The map to be interpolated; a list of vectors.}

\item{oldmap}{Map with positions in the original scale, as in \code{map_to_change}.}

\item{newmap}{Map with positions in the new scale.}
}
\value{
Object of same form as input \code{map} but in the units as in \code{newmap}.
}
\description{
Use interpolate to convert from one map to another
}

