% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_phase_known.R
\name{is_phase_known}
\alias{is_phase_known}
\title{Is cross phase-known}
\usage{
is_phase_known(cross)
}
\arguments{
\item{cross}{Object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}
}
\value{
TRUE/FALSE indicating whether the cross is phase-known or not.
}
\description{
Determine if a cross is of phase-known type
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
is_phase_known(grav2)
}
\keyword{utilities}
