% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_scan1.R
\name{max_scan1}
\alias{max_scan1}
\alias{max.scan1}
\title{Find position with maximum LOD score}
\usage{
max_scan1(scan1_output, lodcolumn = 1, chr = NULL, na.rm = TRUE, ...)

\method{max}{scan1}(scan1_output, lodcolumn = 1, chr = NULL, na.rm = TRUE,
  ...)
}
\arguments{
\item{scan1_output}{An object of class \code{"scan1"} as returned by
\code{\link{scan1}}.}

\item{lodcolumn}{An integer or character string indicating the LOD
score column, either as a numeric index or column name.}

\item{chr}{Option vector of chromosomes to consider.}

\item{na.rm}{Ignored (take to be TRUE)}

\item{...}{Ignored}
}
\value{
A data.frame with three columns: chr, pos, and lod score.
}
\description{
Return data frame with the positions having maximum LOD score for a
particular LOD score column
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# perform genome scan
out <- scan1(probs, pheno, addcovar=covar, Xcovar=Xcovar)

# maximum of first column
max(out)

# maximum of spleen column
max(out, lodcolumn="spleen")

# maximum of first column on chr 2
max(out, chr="2")
}
