% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_scan1.R
\name{maxlod}
\alias{maxlod}
\title{Overall maximum LOD score}
\usage{
maxlod(scan1_output, chr = NULL)
}
\arguments{
\item{scan1_output}{An object of class \code{"scan1"} as returned by
\code{\link{scan1}} or \code{\link{scan1_lmm}}.}

\item{chr}{Option vector of chromosomes to consider.}
}
\value{
A single number: the maximum LOD score across all columns and positions for
the selected chromosomes.
}
\description{
Find overall maximum LOD score in genome scan results.
}
\examples{
\dontrun{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# perform genome scan
out <- scan1(probs, pheno, covar, Xcovar)

# overall maximum
maxlod(out)

# maximum on chromosome 2
max(out, "2")
}
}

