% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmarg.R
\name{maxmarg}
\alias{maxmarg}
\title{Find genotypes with maximum marginal probabilities}
\usage{
maxmarg(probs, minprob = 0.95, chr = NULL, pos = NULL,
  return_char = FALSE, quiet = TRUE, cores = 1)
}
\arguments{
\item{probs}{Genotype probabilities, as calculated from
\code{\link{calc_genoprob}}.}

\item{minprob}{Minimum probability for making a call. If maximum
probability is less then this value, give \code{NA}.}

\item{chr}{If provided (along with \code{pos}), consider only the single specified position.}

\item{pos}{If provided (along with \code{chr}), consider only the single specified position.}

\item{return_char}{If TRUE, return genotype names as character strings.}

\item{quiet}{IF \code{FALSE}, print progress messages.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}
}
\value{
If \code{chr} and \code{pos} are provided, a vector of genotypes is returned

Otherwise, the result is a object like that returned by \code{\link{viterbi}},
A list containing the following:
\itemize{
\item \code{geno} - List of two-dimensional arrays of imputed genotypes,
individuals x positions.
\item \code{map} - The genetic map as a list of vectors of marker positions.
\item \code{grid} - A list of logical vectors, indicating which
    positions correspond to a grid of markers/pseudomarkers. (may be
    absent)
\item \code{crosstype} - The cross type of the input \code{cross}.
\item \code{is_x_chr} - Logical vector indicating whether chromosomes
    are to be treated as the X chromosome or not, from input \code{cross}.
\item \code{sex} - Vector of sexes of the individuals, from input
    \code{cross}.
\item \code{cross_info} - Matrix of cross information for the
    individuals, from input \code{cross}.
\item \code{alleles} - Vector of allele codes, from input
    \code{cross}.
\item \code{alleleprob} - Logical value (\code{FALSE}) that
    indicates whether the probabilities are compressed to allele
    probabilities, as from \code{\link{genoprob_to_alleleprob}}.
\item \code{step} - the value of the \code{step} argument.
\item \code{off_end} - the value of the \code{off_end} argument.
\item \code{stepwidth} - the value of the \code{stepwidth} argument.
\item \code{error_prob} - the value of the \code{error_prob} argument.
\item \code{map_function} - the value of the \code{map_function} argument.
}
}
\description{
For each individual at each position, find the genotype with the maximum marginal probability.
}
\examples{
# load data and calculate genotype probabilities
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
pr <- calc_genoprob(iron, step=0, error_prob=0.002)

# full set of imputed genotypes
ginf <- maxmarg(pr)

# imputed genotypes at a fixed position
g <- maxmarg(pr, chr=8, pos=45.5)

# return genotype names rather than integers
g <- maxmarg(pr, chr=8, pos=45.5, return_char=TRUE)
}
\seealso{
\code{\link{sim_geno}}, \code{\link{viterbi}}
}

