% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ci.R
\name{plot_ci}
\alias{plot_ci}
\title{Effect plot with multiple CIs for different groups}
\usage{
plot_ci(
  est,
  se = NULL,
  lo = NULL,
  hi = NULL,
  SEmult = 2,
  labels = NULL,
  swap_axes = FALSE,
  ...
)
}
\arguments{
\item{est}{Vector of estimates}

\item{se}{Vector of standard errors}

\item{lo}{Vector of lower values for the intervals}

\item{hi}{Vector of upper values for the intervals}

\item{SEmult}{SE multiplier to create intervals}

\item{labels}{Labels for the groups (vector of character strings)}

\item{swap_axes}{If TRUE, have group as y-axis; default (FALSE) has
group on x-axis.}

\item{...}{Optional graphics arguments}
}
\value{
None.
}
\description{
Plot a set of confidence intervals.
}
\details{
Provide either \code{se} or both \code{lo} and \code{hi}. In the case that \code{se} is
used, the intervals will be \code{est} +/- \code{SEmult * se}.

If \code{labels} is not provided, group names are taken from the \code{names(est)}.
If that is also missing, we use capital letters.

You can control the CI line widths with \code{ci_lwd} and the color of
the CI segments with \code{ci_col}. You can control the width of the
segments at the top and bottom with \code{ci_endseg}.
}
\examples{
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))
iron <- iron[,"9"] # only chr 9
map <- insert_pseudomarkers(iron$gmap, step=1)
probs <- calc_genoprob(iron, map, error_prob=0.002)
pr <- pull_genoprobpos(probs, map, chr="9", pos=56.6)
out_fit1 <- fit1(pr, iron$pheno[,"spleen"], se=TRUE, zerosum=FALSE)
plot_ci(out_fit1$coef[1:3], out_fit1$SE[1:3],
        xlab="Genotype chr 9 @ 56.6 cM", ylab="Spleen phenotype")

}
\seealso{
\code{\link[=plot_coef]{plot_coef()}}, \code{\link[=plot_pxg]{plot_pxg()}}, \code{\link[=fit1]{fit1()}}, \code{\link[=pull_genoprobpos]{pull_genoprobpos()}}
}
\keyword{graphics}
