% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot.scan1coef}
\alias{plot_coef}
\alias{plot_coefCC}
\title{Plot QTL effects along chromosome}
\usage{
plot_coef(x, columns, col, add = FALSE, gap = 25, ylim,
  bgcolor = "gray90", altbgcolor = "gray85", ylab = "QTL effects", ...)

plot_coefCC(x, add = FALSE, gap = 25, ylim = NULL, bgcolor = "gray90",
  altbgcolor = "gray85", ylab = "QTL effects", ...)

\method{plot}{scan1coef}(x, columns = 1, col = NULL, add = FALSE,
  gap = 25, ylim = NULL, bgcolor = "gray90", altbgcolor = "gray85",
  ylab = "QTL effects", ...)
}
\arguments{
\item{x}{Estimated QTL effects ("coefficients") as obtained from
\code{\link[qtl2scan]{scan1coef}} or
\code{\link[qtl2scan]{scan1coef_lmm}}.}

\item{columns}{Vector of columns to plot}

\item{col}{Vector of colors, same length as \code{columns}.}

\item{add}{If TRUE, add to current plot (must have same map and
chromosomes).}

\item{gap}{Gap between chromosomes.}

\item{ylim}{y-axis limits}

\item{bgcolor}{Background color for the plot.}

\item{altbgcolor}{Background color for alternate chromosomes.}

\item{ylab}{y-axis label}

\item{...}{Additional graphics parameters.}
}
\description{
Plot estimated QTL effects along a chromosomes.
}
\details{
\code{plot_coefCC()} is the same as \code{plot_coef()}, but forcing
\code{columns=1:8} and using the Collaborative Cross colors,
\code{\link{CCcolors}}.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno[,1]
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)

# calculate coefficients for chromosome 7
library(qtl2scan)
coef <- scan1coef(probs[,7], pheno, covar)

# plot QTL effects
plot(coef, columns=1:3, col=c("slateblue", "violetred", "green3"))
}
\seealso{
\code{\link{CCcolors}}, \code{\link{plot_scan1}}, \code{\link{plot_snpasso}}
}

