% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_geno.R
\name{plot_compare_geno}
\alias{plot_compare_geno}
\alias{plot.compare_geno}
\title{Plot of compare_geno object.}
\usage{
plot_compare_geno(x, rug = TRUE, ...)

\method{plot}{compare_geno}(x, rug = TRUE, ...)
}
\arguments{
\item{x}{A square matrix with genotype comparisons for pairs
of individuals, as output by \code{\link[=compare_geno]{compare_geno()}}.}

\item{rug}{If true, use \code{\link[=rug]{rug()}} to plot tick marks at observed values below the histogram.}

\item{...}{Additional graphics parameters passed to \code{\link[=hist]{hist()}}}
}
\value{
None.
}
\description{
From results of \code{\link[=compare_geno]{compare_geno()}}, plot histogram of
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
cg <- compare_geno(grav2)
plot(cg)
}
\keyword{graphics}
