% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_onegeno.R
\name{plot_onegeno}
\alias{plot_onegeno}
\title{Plot one individual's genome-wide genotypes}
\usage{
plot_onegeno(geno, map, ind = 1, chr = NULL, col = NULL,
  na_col = "white", border = "black", shift = FALSE, bgcolor = "gray90",
  chrwidth = 0.5, ...)
}
\arguments{
\item{geno}{Imputed phase-known genotypes, as a list of matrices
(as produced by \code{\link[qtl2geno]{maxmarg}}) or a list of
three-dimensional arrays (as produced by \code{\link[qtl2geno]{guess_phase}}).}

\item{map}{Marker map (a list of vectors of marker positions).}

\item{ind}{Individual to plot, either a numeric index or an ID.}

\item{chr}{Selected chromosomes to plot; a vector of character strings.}

\item{col}{Vector of colors for the different genotypes.}

\item{na_col}{Color for missing segments.}

\item{border}{Color of outer border around chromosome rectangles.}

\item{shift}{If TRUE, shift the chromosomes so they all start at 0.}

\item{bgcolor}{Color for background rectangle}

\item{chrwidth}{Total width of rectangles for each chromosome, as a
fraction of the distance between them.}

\item{...}{Additional graphics parameters}
}
\description{
Plot one individual's genome-wide genotypes
}
