% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscan1.R
\name{plotscan1}
\alias{plotscan1}
\title{Plot a genome scan}
\usage{
plotscan1(scan1output, lodcolumn = 1, chr, add = FALSE, gap = 25,
  bgcolor = "gray90", altbgcolor = "gray85", ...)
}
\arguments{
\item{scan1output}{Output of \code{\link[qtl2scan]{scan1}} or
\code{\link[qtl2scan]{scan1_lmm}}.}

\item{lodcolumn}{LOD score column to plot.}

\item{chr}{Selected chromosomes to plot; a vector of character
strings.}

\item{add}{If TRUE, add to current plot (must have same map and
chromosomes).}

\item{gap}{Gap between chromosomes.}

\item{bgcolor}{Background color for the plot.}

\item{altbgcolor}{Background color for alternate chromosomes.}

\item{...}{Additional graphics parameters.}
}
\value{
None.
}
\description{
Plot LOD curves for a genome scan
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# perform genome scan
library(qtl2scan)
out <- scan1(probs, pheno, covar, Xcovar)

# plot the results for selected chromosomes
ylim <- c(0, max(out)*1.02)
chr <- c(2,7,8,9,15,16)
plotscan1(out, chr=chr, ylim=ylim)
plotscan1(out, lodcolumn=2, chr=chr, col="violetred", add=TRUE)
legend("topleft", lwd=2, col=c("darkslateblue", "violetred"), colnames(out),
       bg="gray90")

# plot just one chromosome
plotscan1(out, chr=8, ylim=ylim)
plotscan1(out, chr=8, lodcolumn=2, col="violetred", add=TRUE)
}

