% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cross2.R
\name{print.cross2}
\alias{print.cross2}
\title{Print a cross2 object}
\usage{
\method{print}{cross2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}

\item{...}{Ignored.}
}
\value{
None.
}
\description{
Print a summary of a cross2 object
}
\keyword{utilities}

