% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_genoprobpos.R
\name{pull_genoprobpos}
\alias{pull_genoprobpos}
\title{Pull genotype probabilities for a particular position}
\usage{
pull_genoprobpos(genoprobs, marker)
}
\arguments{
\item{genoprobs}{Genotype probabilities as calculated by
\code{\link[=calc_genoprob]{calc_genoprob()}}.}

\item{marker}{A single character string with the name of the
position to pull out.}
}
\value{
A matrix of genotype probabilities for the specified position.
}
\description{
Pull out the genotype probabilities for a particular position (by
name)
}
\examples{
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))
\dontshow{iron <- iron[,"8"]}
gmap <- insert_pseudomarkers(iron$gmap, step=1)
pr <- calc_genoprob(iron, gmap, error_prob=0.002)

pmar <- find_marker(gmap, 8, 40)
pr_8_40 <- pull_genoprobpos(pr, pmar)

}
\seealso{
\code{\link[=find_marker]{find_marker()}}, \code{\link[=fit1]{fit1()}}, \code{\link[=pull_genoprobint]{pull_genoprobint()}}
}
