% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_scan1_map.R, R/qtl2-internal.R
\name{align_scan1_map}
\alias{align_scan1_map}
\alias{qtl2-internal}
\title{Internal functions}
\usage{
align_scan1_map(scan1_output, map)
}
\description{
Functions that are exported and so available to users but not really intended to be used by a typical user.
}
\section{Functions}{
\itemize{
\item \code{align_scan1_map}: Align \code{scan1()} output with a map
}}

\keyword{internal}
