% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2db-package.R
\docType{package}
\name{qtl2db-package}
\alias{qtl2db}
\alias{qtl2db-package}
\title{qtl2db: Connect to Genome Databases for QTL Experiments}
\description{
Functions to connect to genome databases with SNP/variant
and/or gene information, to obtain auxilary information for QTL
analysis. Part of R/qtl2, a reimplementation of the R/qtl package
to better handle high-dimensional data and complex cross designs.
}
\section{Vignettes}{

\itemize{
\item \href{http://kbroman.org/qtl2/assets/vignettes/user_guide.html}{user guide}
\item \href{http://kbroman.org/qtl2/assets/vignettes/input_files.html}{input file formats}
\item \href{http://kbroman.org/qtl2/pages/prep_do_data.html}{preparing DO mouse data for R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/rqtl_diff.html}{differences between R/qtl and R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{developer guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://kbroman.org/qtl2}
  \item \url{https://github.com/rqtl/qtl2db}
  \item Report bugs at \url{https://github.com/rqtl/qtl2db/issues}
}

}
\keyword{internal}
