% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2plot-package.R
\docType{package}
\name{qtl2plot-package}
\alias{qtl2plot}
\alias{qtl2plot-package}
\title{qtl2plot: Data Visualization for QTL Experiments}
\description{
Functions to plot QTL analysis results and related
diagnostics. Part of R/qtl2, a reimplementation of the R/qtl
package to better handle high-dimensional data and complex cross
designs.
}
\section{Vignettes}{

\itemize{
\item \href{http://kbroman.org/qtl2/assets/vignettes/user_guide.html}{user guide}
\item \href{http://kbroman.org/qtl2/assets/vignettes/input_files.html}{input file formats}
\item \href{http://kbroman.org/qtl2/pages/prep_do_data.html}{preparing DO mouse data for R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/rqtl_diff.html}{differences between R/qtl and R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{developer guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://kbroman.org/qtl2}
  \item \url{https://github.com/rqtl/qtl2plot}
  \item Report bugs at \url{https://github.com/rqtl/qtl2plot/issues}
}

}
\keyword{internal}
