% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2version.R
\name{qtl2version}
\alias{qtl2version}
\title{List R/qtl2 package versions.}
\usage{
qtl2version()
}
\value{
Data frame with two columns: package name and version.
}
\description{
List versions of all installed R/qtl2 packages, as shown in the
results of \code{\link[utils:installed.packages]{utils::installed.packages()}}.
}
\examples{
qtl2version()

}
