% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_calc_genoprob.R
\name{rbind.calc_genoprob}
\alias{rbind.calc_genoprob}
\title{Join genotype probabilities for different individuals}
\usage{
\method{rbind}{calc_genoprob}(...)
}
\arguments{
\item{...}{Genotype probability objects as produced by
\code{\link{calc_genoprob}}. Must have the same set of markers and
genotypes.}
}
\value{
A single genotype probability object.
}
\description{
Join multiple genotype probability objects, as produced by
\code{\link{calc_genoprob}} for different individuals.
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
probsA <- calc_genoprob(grav2[1:5,], step=1, error_prob=0.002)
probsB <- calc_genoprob(grav2[6:12,], step=1, error_prob=0.002)
probs <- rbind(probsA, probsB)

}
