% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_phasedgeno.R
\name{rbind.phasedgeno}
\alias{rbind.phasedgeno}
\title{Join phased geno results for different individuals}
\usage{
\method{rbind}{phasedgeno}(...)
}
\arguments{
\item{...}{Imputed genotype objects as produced by
\code{\link[=guess_phase]{guess_phase()}}. Must have the same set of markers.}
}
\value{
An object of class \code{"phasedgeno"}, like the input; see \code{\link[=guess_phase]{guess_phase()}}.
}
\description{
Join phased genotype objects, as produced by \code{\link[=guess_phase]{guess_phase()}},
for the same set of markers but for different individuals.
}
\examples{
\dontrun{
file <- paste0("https://raw.githubusercontent.com/rqtl/",
               "qtl2data/main/DOex/DOex.zip")
DOex <- read_cross2(file)
pr <- calc_genoprob(DOex, error_prob=0.002)
m <- maxmarg(pr)
phA <- guess_phase(DOex[1:10,], m[1:10,])
phB <- guess_phase(DOex[11:20,], m[11:20,])
ph <- rbind(phA, phB)
}

}
\seealso{
\code{\link[=cbind.phasedgeno]{cbind.phasedgeno()}}, \code{\link[=subset.phasedgeno]{subset.phasedgeno()}}, \code{\link[=guess_phase]{guess_phase()}}
}
