% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_cross2}
\alias{read_cross2}
\title{Read QTL data from files}
\usage{
read_cross2(yaml_file)
}
\arguments{
\item{yaml_file}{Character string with path to the yaml file
containing all of the control information.}
}
\value{
Object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer.html}{R/qtl2 developer guide}.
}
\description{
Read QTL data from a set of files
}
\details{
A control file in \href{http://www.yaml.org}{YAML} format contains information about
basic parameters as well as the names of the series of data files
to be read. See the
\href{http://kbroman.org/qtl2/pages/sampledata.html}{sample data files}
and the
\href{http://kbroman.org/qtl2/assets/vignettes/input_files.html}{vignette describing the input file format}.
}
\examples{
yaml_file <- "http://kbroman.org/qtl2/assets/sampledata/grav2/grav2.yaml"
grav2 <- read_cross2(yaml_file)
}
\seealso{
Sample data files at \url{http://kbroman.org/qtl2/pages/sampledata.html}
}
\keyword{IO}

