% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_markers.R
\name{reduce_markers}
\alias{reduce_markers}
\title{Reduce markers to a subset of more-evenly-spaced ones}
\usage{
reduce_markers(map, weights = NULL, min_distance = 1)
}
\arguments{
\item{map}{A list with each component being a vector with the
marker positions for a chromosome.}

\item{weights}{A (optional) list of weights on the markers; same
size as \code{map}.}

\item{min_distance}{Minimum distance between markers.}
}
\value{
A list like the input \code{map}, but with the selected
subset of markers.
}
\description{
Find the largest subset of markers such that no two adjacent
markers are separated by less than some distance.
}
\details{
Uses a dynamic programming algorithm to find, for each
chromosome, the subset of markers for with max(\code{weights}) is
maximal, subject to the constraint that no two adjacent markers may
be separated by more than \code{min_distance}.
}
\examples{
library(qtl)

}
\references{
Broman KW, Weber JL (1999) Method for constructing
confidently ordered linkage maps. Genet Epidemiol 16:337--343
}

