% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1.R
\name{scan1}
\alias{scan1}
\title{Genome scan with a single-QTL model by Haley-Knott regression}
\usage{
scan1(genoprobs, pheno, addcovar = NULL, Xcovar = NULL, intcovar = NULL,
  weights = NULL, cores = 1, ...)
}
\arguments{
\item{genoprobs}{A list of 3-dimensional arrays of genotype
probabilities; each component is a chromosome, and has dimension
individuals x genotypes x positions.}

\item{pheno}{A matrix of phenotypes, individuals x phenotypes.}

\item{addcovar}{An optional matrix of additive covariates.}

\item{Xcovar}{An optional matrix with additional additive covariates used for
null hypothesis when scanning the X chromosome.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{weights}{An optional vector of positive weights for the
individuals. As with the other inputs, it must have \code{names}
for individual identifiers.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}

\item{...}{Additional control parameters; see Details.}
}
\value{
A matrix of LOD scores, positions x phenotypes.  Covariate
column names are included as attributes (\code{"addcovar"},
\code{"intcovar"}, and \code{"Xcovar"}), as is a vector with the
sample size for each phenotype (\code{"sample_size"})
}
\description{
Genome scan with a single-QTL model by Haley-Knott regression, with possible allowance for covariates.
}
\details{
For each of the inputs, the row names are used as
individual identifiers, to align individuals. The \code{genoprobs}
object should have an attribute \code{"is_x_chr"} that indicates
which of the chromosomes is the X chromosome, if any.

The \code{...} argument can contain two additional control
parameters; suspended for simplicity (or confusion, depending on
your point of view). \code{tol} is used as a tolerance value for
linear regression by QR decomposition (in determining whether
columns are linearly dependent on others and should be omitted);
default \code{1e-12}. \code{intcovar_method} indicates whether to
use a high-memory (but potentially faster) method or a low-memory
(and possibly slower) method, with values \code{"highmem"} or
\code{"lowmem"}; default \code{"lowmem"}.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# perform genome scan
out <- scan1(probs, pheno, covar, Xcovar)

}
\references{
Haley CS, Knott SA (1992) A simple
regression method for mapping quantitative trait loci in line
crosses using flanking markers.  Heredity 69:315--324.
}

