% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_lmm.R
\name{scan1_lmm}
\alias{scan1_lmm}
\title{Genome scan with a single-QTL and linear mixed model}
\usage{
scan1_lmm(genoprobs, pheno, kinship, addcovar = NULL, Xcovar = NULL,
  intcovar = NULL, reml = TRUE, cores = 1, ...)
}
\arguments{
\item{genoprobs}{Genotype probabilities as calculated by
\code{\link[qtl2geno]{calc_genoprob}}.}

\item{pheno}{A matrix of phenotypes, individuals x phenotypes.}

\item{kinship}{A kinship matrix, or a list of kinship matrices (one
per chromosome), in order to use the LOCO (leave one chromosome
out) method.}

\item{addcovar}{An optional matrix of additive covariates.}

\item{Xcovar}{An optional matrix with additional additive covariates used for
null hypothesis when scanning the X chromosome.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{reml}{If true, use REML; otherwise, use maximimum likelihood.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}

\item{...}{Additional control parameters; see Details.}
}
\value{
A list containing the following
\itemize{
\item \code{lod} - A matrix of LOD scores, positions x phenotypes.
\item \code{map} - A list containing the map positions at which the
    calculations were performed, taken from the input \code{genoprobs}.
\item \code{hsq} - A matrix of estimated heritabilities under the
    null hypothesis of no QTL. Columns are the phenotypes. If the
    \code{"loco"} method was used with
    \code{\link[qtl2geno]{calc_kinship}} to calculate a list of kinship
    matrices, one per chromosome, the rows of \code{hsq} will be the
    heritabilities for the different chromosomes (well, leaving out
    each one). If \code{Xcovar} was not NULL, there will at least be an
    autosome and X chromosome row.
\item \code{addcovar} - Names of additive covariates that were used.
\item \code{Xcovar} - Names of special covariates for X chromosome
    under the null hypothesis of no QTL
\item \code{intcovar} - Names of interactive covariates that were used.
\item \code{sample_size} - Vector of sample sizes used for each
    phenotype
\item \code{snpinfo} - Present only if the input \code{genoprobs}
    was produced by \code{\link{genoprob_to_snpprob}}, this is a list
    of data frames giving information about all SNPs. The \code{lod}
    matrix will contain only results for distinct SNPs. The
    \code{index} column in \code{snpinfo} is the row index in the
    \code{lod} matrix that corresponds to the current SNP.
}
}
\description{
Genome scan with a single-QTL and linear mixed model to account for
a random polygenic effect, with possible allowance for covariates.
}
\details{
For each of the inputs, the row names are used as
individual identifiers, to align individuals. The \code{genoprobs}
object should have a component \code{"is_x_chr"} that indicates
which of the chromosomes is the X chromosome, if any.

If \code{kinship} is a single matrix, then the \code{hsq}
attribute in the results is a vector of heritabilities. If
\code{kinship} is a list (one matrix per chromosome), then
\code{hsq} is a matrix, phenotypes x chromosomes.

If \code{reml=TRUE}, restricted maximum likelihood (reml) is used
to estimate the heritability under the null hypothesis of no QTL,
separately for each phenotype. But then in the genome scans, this
is taken as fixed and known, and the usual Gaussian log likelihood
is used to calculate LOD scores.

The \code{...} argument can contain two additional control
parameters; suspended for simplicity (or confusion, depending on
your point of view). \code{tol} is used as a tolerance value for
linear regression by QR decomposition (in determining whether
columns are linearly dependent on others and should be omitted);
default \code{1e-12}. \code{intcovar_method} indicates whether to
use a high-memory (but potentially faster) method or a low-memory
(and possibly slower) method, with values \code{"highmem"} or
\code{"lowmem"}; default \code{"lowmem"}.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# leave-one-chromosome-out kinship matrices
kinship <- calc_kinship(probs, "loco")

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# perform genome scan
out <- scan1_lmm(probs, pheno, kinship, covar, Xcovar)

}

