% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1coef_lmm.R
\name{scan1coef_lmm}
\alias{scan1coef_lmm}
\title{Calculate QTL effects in scan along one chromosome, adjusting for polygenes with an LMM}
\usage{
scan1coef_lmm(genoprobs, pheno, kinship, addcovar = NULL, intcovar = NULL,
  contrasts = NULL, se = FALSE, hsq, reml = TRUE, tol = 0.000000000001)
}
\arguments{
\item{genoprobs}{Genotype probabilities as calculated by
\code{\link[qtl2geno]{calc_genoprob}}.}

\item{pheno}{A numeric vector of phenotype values (just one phenotype, not a matrix of them)}

\item{kinship}{A kinship matrix. Can be eigen decomposition of
kinship matrix (as calculated with \code{\link{decomp_kinship}}),
but it would need to be for the exact subset of individuals
(following omitting those with missing genotype probabilities,
phenotypes, or covariates).}

\item{addcovar}{An optional matrix of additive covariates.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{contrasts}{An optional matrix of genotype contrasts, size
genotypes x genotypes. For an intercross, you might use
\code{cbind(c(1,0,0), c(-0.5, 0, 0.5), c(-0.5, 1, 0.5))} to get
mean, additive effect, and dominance effect. The default is the
identity matrix.}

\item{se}{If TRUE, also calculate the standard errors.}

\item{hsq}{(Optional) residual heritability}

\item{reml}{If true and \code{hsq} is not provided, use REML to estimate \code{hsq}.}

\item{tol}{Tolerance value for
linear regression by QR decomposition (in determining whether
columns are linearly dependent on others and should be omitted)}
}
\value{
A matrix of estimated regression coefficients, of dimension
positions x number of effects. The number of effects is
\code{n_genotypes + n_addcovar + (n_genotypes-1)*n_intcovar}. The
map of positions at which the calculations were performed is
included as an attribute \code{"map"} (taken from the
input \code{genoprobs}).
}
\description{
Calculate QTL effects in scan along one chromosome with a
single-QTL model using a linear mixed model, with possible
allowance for covariates.
}
\details{
For each of the inputs, the row names are used as
individual identifiers, to align individuals.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
\dontshow{iron <- iron[,c(7, 8)]}

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# leave-one-chromosome-out kinship matrices
kinship <- calc_kinship(probs, "loco")

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno[,1]
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)

# calculate coefficients for chromosome 7
coef <- scan1coef_lmm(probs[,"7"], pheno, kinship[["7"]], addcovar=covar)

}

