% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.cross2}
\alias{[.cross2}
\alias{subset.cross2}
\title{Subsetting data for a QTL experiment}
\usage{
\method{subset}{cross2}(x, ind, chr, ...)

\method{[}{cross2}(x, ind, chr)
}
\arguments{
\item{x}{An object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}

\item{ind}{A vector of individuals: numeric indices, logical
values, or character string IDs}

\item{chr}{A vector of chromosomes: numeric indices, logical
values, or character string IDs}

\item{...}{Ignored.}
}
\value{
The input \code{cross2} object, with the selected
individuals and/or chromsomes.
}
\description{
Pull out a specified set of individuals and/or chromosomes from a
\code{cross2} object.
}
\section{Warning}{

The order of the two arguments is reversed relative to the related
function in \href{http://www.rqtl.org}{R/qtl}.
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
# keep individuals 1-20 and chromosomes 3 and 4
grav2sub <- grav2[1:20, c(3,4)]
# keep just chromosome 1
grav2_c1 <- grav2[,1]
}
\keyword{utilities}

