% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_phasedgeno.R
\name{subset.phasedgeno}
\alias{subset.phasedgeno}
\alias{[.phasedgeno}
\title{Subsetting phased genotype objects}
\usage{
\method{subset}{phasedgeno}(x, ind = NULL, chr = NULL, ...)

\method{[}{phasedgeno}(x, ind = NULL, chr = NULL)
}
\arguments{
\item{x}{Imputed genotypes as output from \code{\link[=guess_phase]{guess_phase()}}, a list of 3d arrays.}

\item{ind}{A vector of individuals: numeric indices, logical
values, or character string IDs}

\item{chr}{A vector of chromosomes: logical values, or character
string IDs. Numbers are interpreted as character string IDs.}

\item{...}{Ignored.}
}
\value{
An object of class \code{"phasedgeno"}, like the input, with the
selected individuals and/or chromosomes; see \code{\link[=guess_phase]{guess_phase()}}.
}
\description{
Pull out a specified set of individuals and/or chromosomes from
the results of \code{\link[=guess_phase]{guess_phase()}}
}
\examples{
\dontrun{
file <- paste0("https://raw.githubusercontent.com/rqtl/",
               "qtl2data/main/DOex/DOex.zip")
DOex <- read_cross2(file)
pr <- calc_genoprob(DOex, error_prob=0.002)
m <- maxmarg(pr)
ph <- guess_phase(DOex, m)
ph <- ph[1:10,] # first ten individuals
ph <- ph[,2]  # chromosome 2
}
}
\seealso{
\code{\link[=guess_phase]{guess_phase()}}, \code{\link[=cbind.phasedgeno]{cbind.phasedgeno()}}, \code{\link[=rbind.phasedgeno]{rbind.phasedgeno()}}
}
\keyword{utilities}
