% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cross2.R
\name{summary.cross2}
\alias{summary.cross2}
\title{Summary of cross2 object}
\usage{
\method{summary}{cross2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"cross2"}. For details, see the
\href{https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}

\item{...}{Ignored.}
}
\value{
None.
}
\description{
Summarize a cross2 object
}
\seealso{
\link{basic_summaries}
}
\keyword{utilities}
