% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_snps.R
\name{top_snps}
\alias{top_snps}
\title{Create table of top snp associations}
\usage{
top_snps(scan1_output, drop = 1.5, show_all_snps = TRUE)
}
\arguments{
\item{scan1_output}{Output of \code{\link[qtl2scan]{scan1}} or
\code{\link[qtl2scan]{scan1_lmm}}. Should contain an attribute,
\code{"snpinfo"}, as when \code{\link[qtl2scan]{scan1}} or
\code{\link[qtl2scan]{scan1_lmm}} are run with SNP probabilities
produced by \code{\link[qtl2scan]{genoprob_to_snpprob}}.}

\item{drop}{Show all SNPs with LOD score within this amount of the
maximum SNP association.}

\item{show_all_snps}{If TRUE, expand to show all SNPs.}
}
\description{
Create a table of the top snp associations
}
\seealso{
\code{\link{genoprob_to_snpprob}}, \code{\link[qtl2plot]{plot_snpasso}}
}

