% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/calc_genetic_sim.R
\name{calc_genetic_sim}
\alias{calc_genetic_sim}
\title{Calculate genetic similarity among individuals}
\usage{
calc_genetic_sim(probs, use_grid_only = TRUE, omit_x = TRUE, quiet = TRUE)
}
\arguments{
\item{probs}{List of three-dimensional arrays of probabilities, as calculated from \code{\link{calc_genoprob}}.}

\item{use_grid_only}{If \code{TRUE} and \code{probs} were calculated with
\code{stepwidth="fixed"}, reduce them to the grid using
\code{\link{probs_to_grid}}.}

\item{omit_x}{If \code{TRUE}, only use the autosomes.}

\item{quiet}{IF \code{FALSE}, print messages on progress.}
}
\value{
A matrix of proportion of matching genotypes.
}
\description{
Calculate genetic similarity among individuals from conditional genotype probabilities.
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
probs <- calc_genoprob(grav2, step=1, error_prob=0.002)
sim <- calc_genetic_sim(probs)
}
\keyword{utilities}

