% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_scan1.R
\name{cbind.scan1}
\alias{cbind.scan1}
\title{Join genome scan results for different phenotypes.}
\usage{
\method{cbind}{scan1}(...)
}
\arguments{
\item{...}{Genome scan objects as produced by \code{\link{scan1}}.
Must have the same map.}
}
\value{
A single genome scan object with the lod score columns
combined as multiple columns.
}
\description{
Join multiple \code{\link{scan1}} results for different phenotypes;
must have the same map.
}
\details{
If components \code{addcovar}, \code{Xcovar},
\code{intcovar}, \code{weights} do not match between objects, we
omit this information.

If \code{hsq} present but has differing numbers of rows, we omit this information.

If \code{snpinfo} present, they much match.
}
\examples{
library(qtl2geno)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probs <- calc_genoprob(grav2, map, error_prob=0.002)
phe1 <- grav2$pheno[,1,drop=FALSE]
phe2 <- grav2$pheno[,2,drop=FALSE]

out1 <- scan1(probs, phe1) # phenotype 1
out2 <- scan1(probs, phe2) # phenotype 2
out <- cbind(out1, out2)

}
