% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCcolors-data.R
\name{CCcolors}
\alias{CCcolors}
\alias{CCorigcolors}
\alias{CCaltcolors}
\title{Collaborative Cross colors}
\source{
\url{https://csbio.unc.edu/CCstatus/index.py?run=AvailableLines.information}
}
\description{
A vector of 8 colors for use with the mouse
Collaborative Cross and Diversity Outbreds.
}
\details{
\code{CCorigcolors} are the original eight colors for the Collaborative Cross founder strains.
\code{CCaltcolors} are a slightly modified version, but still not color-blind friendly.
\code{CCcolors} are derived from the the Okabe-Ito color blind friendly palette in
\href{https://doi.org/10.1038/nmeth.1618}{Wong (2011) Nature Methods}.
}
\examples{
data(CCcolors)
data(CCaltcolors)
data(CCorigcolors)
}
\keyword{datasets}
