% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCcolors-data.R
\name{CCcolors}
\alias{CCcolors}
\alias{CCorigcolors}
\title{Collaborative Cross colors}
\source{
\url{http://compgen.unc.edu/wp/?page_id=577}
}
\description{
A vector of 8 colors for use with the mouse
Collaborative Cross and Diversity Outbreds.
}
\details{
\code{CCorigcolors} are the original eight colors for the Collaborative Cross founder strains.
\code{CCcolors} are slightly modified.
}
\examples{
data(CCcolors)
data(CCorigcolors)
}
\keyword{datasets}
