% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert2cross2}
\alias{convert2cross2}
\title{Convert R/qtl cross object to new format}
\usage{
convert2cross2(cross)
}
\arguments{
\item{cross}{An object of class \code{"cross"}; see
\code{\link[qtl]{read.cross}} for details.}
}
\value{
Object of class \code{"cross2"}. For details, see the
R/qtl2 Developer Guide. **FIX ME (put the url here)**
}
\description{
Convert a cross object from the R/qtl format to the R/qtl2 format
}
\examples{
library(qtl)
data(hyper)
hyper2 <- convert2cross2(hyper)
}
\seealso{
\code{\link{read_cross2}}
}
\keyword{utilities}

