% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2scanone.R
\name{convert2scanone}
\alias{convert2scanone}
\title{Convert scan1 results to the scanone format}
\usage{
convert2scanone(output, probs, map)
}
\arguments{
\item{output}{Matrix of LOD scores, as calculated by \code{\link{scan1}}.}

\item{probs}{Genotype probabilities (calculated with the qtl2geno
function \code{calc_genoprob}) used to calculate \code{output}. It
includes, as an attribute, the map of positions at which the LOD
scores were calculated.}

\item{map}{Genetic map of positions at which the LOD scores were
calculated. (Needed if \code{probs} is not provided.)}
}
\value{
A data frame with class \code{"scanone"}, containing
chromosome and position columns followed by the LOD scores in
\code{output}.
}
\description{
Convert the results of \code{\link{scan1}} to the form used by the
R/qtl function \code{scanone}.
}
\details{
Provide either \code{map} or \code{probs}; the latter
contains the map as an attribute.
}
\examples{
library(qtl2geno)
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
probs <- calc_genoprob(iron, step=1, error_prob=0.002)
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)
out <- scan1(probs, pheno, covar, Xcovar)

out_rev <- convert2scanone(out, probs)

}

