% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variant_query_func.R
\name{create_variant_query_func}
\alias{create_variant_query_func}
\title{Create a function to query variants}
\usage{
create_variant_query_func(dbfile = NULL, db = NULL,
  table_name = "variants", chr_field = "chr", pos_field = "pos_Mbp")
}
\arguments{
\item{dbfile}{Name of database file}

\item{db}{Optional database connection (provide one of \code{file} and \code{db}).}

\item{table_name}{Name of table in the database}

\item{chr_field}{Name of chromosome field}

\item{pos_field}{Name of position field}
}
\value{
Function with three arguments, \code{chr}, \code{start}, and \code{end},
which returns a data frame with the variants in that region. The output should contain
at least the columns \code{chr} and \code{pos}, the latter being position in Mbp.
}
\description{
Create a function that will connect to a SQLite database of founder
variant information and return a data frame with variants for a
selected region.
}
\examples{
\dontrun{
# create query function by connecting to file
query_variants <- create_variant_query_func("cc_variants.sqlite")
# query_variants will connect and disconnect each time
variants <- query_variants("19", 25.1, 26.1)

# connect and disconnect separately
library(RSQLite)
db <- dbConnect(SQLite(), "cc_variants.sqlite")
query_variants <- create_variant_query_func(db=db)
variants <- query_variants("19", 25.1, 26.1)
dbDisconnect(db)
}
}
