% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ibd_segments.R
\name{find_ibd_segments}
\alias{find_ibd_segments}
\title{Find IBD segments for a set of strains}
\usage{
find_ibd_segments(geno, map, min_lod = 15, error_prob = 0.001, cores = 1)
}
\arguments{
\item{geno}{List of matrices of founder genotypes. The matrices
correspond to the genotypes on chromosomes and are arrayed as
founders x markers.}

\item{map}{List of vectors of marker positions}

\item{min_lod}{Threshold for minimum LOD score for a segment}

\item{error_prob}{Genotyping error/mutation probability}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}
}
\value{
A data frame whose rows are IBD segments and whose columns
    are:
\itemize{
\item Strain 1
\item Strain 2
\item Chromosome
\item Left marker
\item Right marker
\item Left position
\item Right position
\item Interval length
\item Number of markers
\item Number of mismatches
\item LOD score
}
}
\description{
Find IBD segments (regions with a lot of shared SNP genotypes) for
a set of strains
}
\examples{
\donttest{
# load DO data from Recla et al. (2014) Mamm Genome 25:211-222.
recla <- read_cross2("https://raw.githubusercontent.com/rqtl/qtl2data/master/DO_Recla/recla.zip")

# grab founder genotypes and physical map
fg <- recla$founder_geno
pmap <- recla$pmap

# find shared segments
(segs <- find_ibd_segments(fg, pmap, min_lod=10, error_prob=0.0001))
}

}
\references{
Broman KW, Weber JL (1999) Long homozygous chromosomal segments in reference families from the
Centre d’Etude du Polymorphisme Humain. Am J Hum Genet 65:1493-1500.
}
