% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit1.R
\name{fit1}
\alias{fit1}
\title{Fit single-QTL model at a single position}
\usage{
fit1(genoprobs, pheno, kinship = NULL, addcovar = NULL, nullcovar = NULL,
  intcovar = NULL, weights = NULL, contrasts = NULL, model = c("normal",
  "binary"), se = TRUE, hsq = NULL, reml = TRUE, tol = 0.000000000001,
  maxit = 100)
}
\arguments{
\item{genoprobs}{A matrix of genotype probabilities, individuals x genotypes}

\item{pheno}{A numeric vector of phenotype values (just one phenotype, not a matrix of them)}

\item{kinship}{Optional kinship matrix.}

\item{addcovar}{An optional matrix of additive covariates.}

\item{nullcovar}{An optional matrix of additional additive
covariates that are used under the null hypothesis (of no QTL)
but not under the alternative (with a QTL). This is needed for
the X chromosome, where we might need sex as a additive
covariate under the null hypothesis, but we wouldn't want to
include it under the alternative as it would be collinear with
the QTL effects.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{weights}{An optional vector of positive weights for the
individuals. As with the other inputs, it must have \code{names}
for individual identifiers. Ignored if \code{kinship} is provided.}

\item{contrasts}{An optional matrix of genotype contrasts, size
genotypes x genotypes. For an intercross, you might use
\code{cbind(c(1,1,1), c(-1, 0, 1), c(-0.5, 1, -0.5))} to get
mean, additive effect, and dominance effect. The default is the
identity matrix.}

\item{model}{Indicates whether to use a normal model (least
squares) or binary model (logistic regression) for the phenotype.
If \code{model="binary"}, the phenotypes must have values in
\eqn{[0, 1]}.}

\item{se}{If TRUE, calculate the standard errors.}

\item{hsq}{(Optional) residual heritability; used only if
\code{kinship} provided.}

\item{reml}{If \code{kinship} provided: if \code{reml=TRUE}, use
REML; otherwise maximum likelihood.}

\item{tol}{Tolerance value for
linear regression by QR decomposition (in determining whether
columns are linearly dependent on others and should be omitted)}

\item{maxit}{Maximum number of iterations in logistic regression
fit (when \code{model="binary"}).}
}
\value{
A list containing
\itemize{
\item \code{coef} - Vector of estimated coefficients.
\item \code{SE} - Vector of estimated standard errors (included if \code{se=TRUE}).
\item \code{lod} - The overall lod score.
\item \code{ind_lod} - Vector of individual contributions to the LOD score.
\item \code{fitted}  - Fitted values.
}
}
\description{
Fit a single-QTL model at a single putative QTL position and get detailed results
about estimated coefficients and individuals contributions to the LOD score.
}
\details{
For each of the inputs, the row names are used as
individual identifiers, to align individuals.

If \code{kinship} is absent, Haley-Knott regression is performed.
If \code{kinship} is provided, a linear mixed model is used, with a
polygenic effect estimated under the null hypothesis of no (major)
QTL, and then taken as fixed as known in the genome scan.
}
\examples{
# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))

# insert pseudomarkers into map
map <- insert_pseudomarkers(iron$gmap, step=1)

# calculate genotype probabilities
probs <- calc_genoprob(iron, map, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno[,1]
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)

# leave-one-chromosome-out kinship matrix for chr 7
kinship7 <- calc_kinship(probs, "loco")[[7]]

# scan chromosome 7 to find peak
out <- scan1(probs[,7], pheno, kinship7, addcovar=covar)

# find peak position
max_pos <- rownames(max(out, map[7]))

# fit QTL model just at that position
out_fit1 <- fit1(probs[[7]][,,max_pos], pheno, addcovar=covar)

# fit QTL model just at that position, with polygenic effect
out_fit1_pg <- fit1(probs[[7]][,,max_pos], pheno, kinship7, addcovar=covar)

}
\references{
Haley CS, Knott SA (1992) A simple
regression method for mapping quantitative trait loci in line
crosses using flanking markers.  Heredity 69:315--324.

Kang HM, Zaitlen NA, Wade CM, Kirby A, Heckerman D, Daly MJ, Eskin
E (2008) Efficient control of population structure in model
organism association mapping. Genetics 178:1709--1723.
}
\seealso{
\code{\link[=pull_genoprobpos]{pull_genoprobpos()}}
}
