% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoprob_to_alleleprob.R
\name{genoprob_to_alleleprob}
\alias{genoprob_to_alleleprob}
\title{Convert genotype probabilities to allele probabilities}
\usage{
genoprob_to_alleleprob(probs, quiet = TRUE, n_cores = 1)
}
\arguments{
\item{probs}{List of three-dimensional arrays of probabilities, as
calculated from \code{\link{calc_genoprob}}.}

\item{quiet}{IF \code{FALSE}, print progress messages.}

\item{n_cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)}
}
\value{
List of three-dimensional arrays of probabilities,
regarding alleles rather than genotypes.
}
\description{
Reduce genotype probabilities (as calculated by
\code{\link{calc_genoprob}}) to allele probabilities.
}
\examples{
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
probs <- calc_genoprob(iron, step=1, error_prob=0.002)
allele_probs <- genoprob_to_alleleprob(probs)
}
\keyword{utilities}

