% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cistrans.R
\name{plot_cistrans}
\alias{plot_cistrans}
\title{cis-trans plot for eQTL results}
\usage{
plot_cistrans(
  peaks,
  map,
  pheno_pos,
  gap = 0,
  pty = "s",
  bgcolor = "gray90",
  altbgcolor = "gray80",
  ...
)
}
\arguments{
\item{peaks}{Data frame of QTL results, as output by \code{\link[=find_peaks]{find_peaks()}}
Should contain columns \code{lodcolumn}, \code{chr}, and \code{pos}.}

\item{map}{Marker map, as a list of chromosomes, each being a
vector of positions; used to get chromosome names and start/end}

\item{pheno_pos}{Data frame with chromosome, position, and name of
the phenotypes in \code{peaks}. Should contain columns \code{pheno},
\code{chr}, and \code{pos}.}

\item{gap}{Gap between chromosomes in the plot}

\item{pty}{Plot type (generally "s" to force a square plot)}

\item{bgcolor}{Background color for chromosome rectangles}

\item{altbgcolor}{Alternate background color for chromosome rectangles}

\item{...}{Additional graphics arguments passed to \code{\link[graphics:points]{graphics::points()}}}
}
\value{
None
}
\description{
Scatterplot of gene location vs QTL location to display eQTL/pQTL results
}
\examples{
\dontrun{
# download example pQTL results (from Keele et al. 2026, https://doi.org/10.1016/j.xgen.2025.101069)
# contains qtl, map_endpoints, and pheno_pos
download.file("https://kbroman.org/qtl2/assets/sampledata/pqtl_data.RData", "pqtl_data.RData")
load("pqtl_data.RData")

# adjust pheno_pos
pheno_pos$pheno <- pheno_pos$uniprot_id
pheno_pos$pos <- (pheno_pos$start + pheno_pos$end)/2

plot_cistrans(qtl, map_endpoints, pheno_pos)
}
}
\seealso{
\code{\link[=plot_scan1]{plot_scan1()}}, \code{\link[=find_peaks]{find_peaks()}}, \code{\link[=plot_lodpeaks]{plot_lodpeaks()}}
}
