% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genoprob.R
\name{plot_genoprob}
\alias{plot_genoprob}
\title{Plot genotype probabilities for one individual on one chromosome.}
\usage{
plot_genoprob(probs, map, ind = 1, chr = NULL, geno = NULL,
  color_scheme = c("gray", "viridis"), col = NULL, threshold = 0,
  transpose = FALSE, hlines = NULL, hlines.col = "#B3B3B370",
  hlines.lwd = 1, hlines.lty = 1, vlines = NULL,
  vlines.col = "#B3B3B370", vlines.lwd = 1, vlines.lty = 1, ...)
}
\arguments{
\item{probs}{Genotype probabilities (as produced by \code{\link[qtl2geno]{calc_genoprob}})
or allele dosages (as produced by \code{\link[qtl2geno]{genoprob_to_alleleprob}}).}

\item{map}{Marker map (a list of vectors of marker positions).}

\item{ind}{Individual to plot, either a numeric index or an ID.}

\item{chr}{Selected chromosome to plot; a single character string.}

\item{geno}{Optional vector of genotypes or alleles to be shown.}

\item{col}{Optional vector of colors for the heatmap.}

\item{threshold}{Threshold for genotype probabilities; only genotypes that achieve}

\item{transpose}{If TRUE, swap the axes, so that the genotypes are
on the x-axis and the chromosome position is on the y-axis.}

\item{hlines}{Position of horizontal grid lines (use \code{NA} to avoid lines).}

\item{hlines.col}{Color of horizontal grid lines.}

\item{hlines.lwd}{Line width of horizontal grid lines.}

\item{hlines.lty}{Line type of horizontal grid lines.}

\item{vlines}{Position of vertical grid lines (use \code{NA} to avoid lines).}

\item{vlines.col}{Color of vertical grid lines.}

\item{vlines.lwd}{Line width of vertical grid lines.}

\item{vlines.lty}{Line type of vertical grid lines.

this value somewhere on the chromosome will be shown.}

\item{...}{Additional graphics parameters passed to \code{\link[graphics]{image}}.}

\item{colscheme}{Color scheme for the heatmap (ignored if \code{col} is provided).}
}
\description{
Plot the genotype probabilities for one individual on one chromosome, as a heat map.
}
