% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snpasso.R
\name{plot_snpasso}
\alias{plot_snpasso}
\title{Plot SNP associations}
\usage{
plot_snpasso(scan1output, show_all_snps = TRUE, drop.hilit = NA,
  col.hilit = "violetred", col = "darkslateblue", pch = 16, cex = 0.5,
  ylim = NULL, add = FALSE, gap = 25, bgcolor = "gray90",
  altbgcolor = "gray85", ...)
}
\arguments{
\item{scan1output}{Output of \code{\link[qtl2scan]{scan1}} or
\code{\link[qtl2scan]{scan1_lmm}}. Should contain an attribute,
\code{"snpinfo"}, as when \code{\link[qtl2scan]{scan1}} or
\code{\link[qtl2scan]{scan1_lmm}} are run with SNP probabilities
produced by \code{\link[qtl2scan]{genoprob_to_snpprob}}.}

\item{show_all_snps}{If TRUE, expand to show all SNPs.}

\item{drop.hilit}{SNPs with LOD score within this amount of the maximum SNP association will be highlighted.}

\item{col.hilit}{Color of highlighted points}

\item{col}{Color of other points}

\item{pch}{Plotting character for the points (default 16)}

\item{cex}{Character expansion for the points (default 0.5)}

\item{ylim}{y-axis limits}

\item{add}{If TRUE, add to current plot (must have same map and
chromosomes).}

\item{gap}{Gap between chromosomes.}

\item{bgcolor}{Background color for the plot.}

\item{altbgcolor}{Background color for alternate chromosomes.}

\item{...}{Additional graphics parameters.}
}
\description{
Plot SNP associations, with possible expansion from distinct snps to all snps.
}
\seealso{
\code{\link{plot_scan1}}, \code{\link{plot_coef}}, \code{\link{plot_coefCC}}
}

