% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probs_to_grid.R
\name{probs_to_grid}
\alias{probs_to_grid}
\title{Subset genotype probability array to pseudomarkers on a grid}
\usage{
probs_to_grid(probs)
}
\arguments{
\item{probs}{List of 3d arrays, as output from
\code{\link{calc_genoprob}} with \code{stepwidth="fixed"}.}
}
\value{
Same list as input, but subset to just include
pseudomarkers along a grid.
}
\description{
Subset genotype probability array (from \code{\link{calc_genoprob}}
to a grid of pseudomarkers along each chromosome.
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
probs <- calc_genoprob(grav2, step=1, error_prob=0.002)
sapply(probs, dim)
probs_sub <- probs_to_grid(probs)
sapply(probs_sub, dim)
}
\keyword{utilities}

