% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2.R
\docType{package}
\name{qtl2-package}
\alias{qtl2}
\alias{qtl2-package}
\title{qtl2: Easily Install and Load R/qtl2}
\description{
R/qtl2 is a set of packages that are used together to
perform QTL analysis. It is a reimplementation of the R/qtl
package to better handle high-dimensional data and complex cross
designs. This package is designed to make it easy to install and
load multiple R/qtl2 packages in a single step.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://kbroman.org/qtl2}
  \item \url{https://github.com/rqtl/qtl2}
  \item Report bugs at \url{https://github.com/rqtl/qtl2/issues}
}

}
\keyword{internal}
