% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2-package.R
\docType{package}
\name{qtl2-package}
\alias{qtl2}
\alias{qtl2-package}
\title{qtl2: Quantitative Trait Locus Mapping in Experimental Crosses}
\description{
R/qtl2 provides a set of tools to perform quantitative
    trait locus (QTL) analysis in experimental crosses. It is a
    reimplementation of the R/qtl package to better handle
    high-dimensional data and complex cross designs. This package is
    designed to make it easy to install and load multiple R/qtl2
    packages in a single step.
}
\section{Vignettes}{

\itemize{
\item \href{http://kbroman.org/qtl2/assets/vignettes/user_guide.html}{user guide}
\item \href{http://kbroman.org/qtl2/assets/vignettes/input_files.html}{input file formats}
\item \href{http://kbroman.org/qtl2/pages/prep_do_data.html}{preparing DO mouse data for R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/rqtl_diff.html}{differences between R/qtl and R/qtl2}
\item \href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{developer guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://kbroman.org/qtl2}
  \item \url{https://github.com/rqtl/qtl2}
  \item Report bugs at \url{https://github.com/rqtl/qtl2/issues}
}

}
\keyword{internal}
