% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2version.R
\name{qtl2version}
\alias{qtl2version}
\title{Installed version of R/qtl2}
\usage{
qtl2version()
}
\value{
A character string with the installed version of the R/qtl2 package.
}
\description{
Get installed version of R/qtl2
}
\examples{
qtl2version()

}
