% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_cross2}
\alias{read_cross2}
\title{Read QTL data from files}
\usage{
read_cross2(yaml_file)
}
\arguments{
\item{yaml_file}{Character string with path to the yaml file
containing all of the control information.}
}
\value{
Object of class \code{"cross2"}. For details, see the
R/qtl2 Developer Guide. **FIX ME (put the url here)**
}
\description{
Read QTL data from a set of files
}
\details{
A control file in YAML format contains information above
basic parameters as well as the names of the series of data files
to be read. Sample data files at **FIX ME (put url here)**.
}
\examples{
yaml_file <- "http://kbroman.org/qtl2/assets/sampledata/grav2/grav2.yaml"
grav2 <- read_cross2(yaml_file)
}
\seealso{
Sample data files at \url{http://kbroman.org/qtl2/pages/sampledata.html}
}
\keyword{IO}

