% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1.R
\name{scan1}
\alias{scan1}
\title{Genome scan with a single-QTL model}
\usage{
scan1(genoprobs, pheno, kinship = NULL, addcovar = NULL, Xcovar = NULL,
  intcovar = NULL, weights = NULL, reml = TRUE, cores = 1, ...)
}
\arguments{
\item{genoprobs}{Genotype probabilities as calculated by
\code{\link[qtl2geno]{calc_genoprob}}.}

\item{pheno}{A matrix of phenotypes, individuals x phenotypes.}

\item{kinship}{Optional kinship matrix, or a list of kinship matrices (one
per chromosome), in order to use the LOCO (leave one chromosome
out) method.}

\item{addcovar}{An optional matrix of additive covariates.}

\item{Xcovar}{An optional matrix with additional additive covariates used for
null hypothesis when scanning the X chromosome.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{weights}{An optional vector of positive weights for the
individuals. As with the other inputs, it must have \code{names}
for individual identifiers. Ignored if \code{kinship} is provided.}

\item{reml}{If \code{kinship} provided: if \code{reml=TRUE}, use
REML; otherwise maximum likelihood.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}

\item{...}{Additional control parameters; see Details.}
}
\value{
A list containing the following
\itemize{
\item \code{lod} - A matrix of LOD scores, positions x phenotypes.
\item \code{map} - A list containing the map positions at which the
    calculations were performed, taken from the input \code{genoprobs}.
\item \code{addcovar} - Names of additive covariates that were used.
\item \code{Xcovar} - Names of special covariates for X chromosome
    under the null hypothesis of no QTL
\item \code{intcovar} - Names of interactive covariates that were used.
\item \code{sample_size} - Vector of sample sizes used for each
    phenotype
\item \code{weights} - Logical value indicating whether weights
    were used (only included if \code{kinship} was not).
\item \code{hsq} - Included if \code{kinship} provided: A matrix of
    estimated heritabilities under the null hypothesis of no
    QTL. Columns are the phenotypes. If the \code{"loco"} method was
    used with \code{\link[qtl2geno]{calc_kinship}} to calculate a list
    of kinship matrices, one per chromosome, the rows of \code{hsq}
    will be the heritabilities for the different chromosomes (well,
    leaving out each one). If \code{Xcovar} was not NULL, there will at
    least be an autosome and X chromosome row.
\item \code{snpinfo} - Present only if the input \code{genoprobs}
    was produced by \code{\link{genoprob_to_snpprob}}, this is a list
    of data frames giving information about all SNPs. The \code{lod}
    matrix will contain only results for distinct SNPs. The
    \code{index} column in \code{snpinfo} is the row index in the
    \code{lod} matrix that corresponds to the current SNP.
\item \code{is_x_chr} - Taken from \code{genoprobs}; indicates
    which chromosomes are the X chromosome.
}
}
\description{
Genome scan with a single-QTL model by Haley-Knott regression or a
linear mixed model, with possible allowance for covariates.
}
\details{
For each of the inputs, the row names are used as
individual identifiers, to align individuals. The \code{genoprobs}
object should have a component \code{"is_x_chr"} that indicates
which of the chromosomes is the X chromosome, if any.

The \code{...} argument can contain two additional control
parameters; suspended for simplicity (or confusion, depending on
your point of view). \code{tol} is used as a tolerance value for
linear regression by QR decomposition (in determining whether
columns are linearly dependent on others and should be omitted);
default \code{1e-12}. \code{intcovar_method} indicates whether to
use a high-memory (but potentially faster) method or a low-memory
(and possibly slower) method, with values \code{"highmem"} or
\code{"lowmem"}; default \code{"lowmem"}.

If \code{kinship} is absent, Haley-Knott regression is performed.
If \code{kinship} is provided, a linear mixed model is used, with a
polygenic effect estimated under the null hypothesis of no (major)
QTL, and then taken as fixed as known in the genome scan.

If \code{kinship} is a single matrix, then the \code{hsq}
in the results is a vector of heritabilities (one value for each phenotype). If
\code{kinship} is a list (one matrix per chromosome), then
\code{hsq} is a matrix, chromosomes x phenotypes.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# perform genome scan
out <- scan1(probs, pheno, addcovar=covar, Xcovar=Xcovar)

# leave-one-chromosome-out kinship matrices
kinship <- calc_kinship(probs, "loco")

# genome scan with a linear mixed model
out_lmm <- scan1(probs, pheno, kinship, covar, Xcovar)

}
\references{
Haley CS, Knott SA (1992) A simple
regression method for mapping quantitative trait loci in line
crosses using flanking markers.  Heredity 69:315--324.

Kang HM, Zaitlen NA, Wade CM, Kirby A, Heckerman D, Daly MJ, Eskin
E (2008) Efficient control of population structure in model
organism association mapping. Genetics 178:1709--1723.
}

