% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1coef.R
\name{scan1coef}
\alias{scan1coef}
\title{Calculate QTL effects in scan along one chromosome}
\usage{
scan1coef(genoprobs, pheno, kinship = NULL, addcovar = NULL,
  intcovar = NULL, weights = NULL, contrasts = NULL, se = FALSE,
  hsq = NULL, reml = TRUE, tol = 0.000000000001)
}
\arguments{
\item{genoprobs}{Genotype probabilities as calculated by
\code{\link[qtl2geno]{calc_genoprob}}.}

\item{pheno}{A numeric vector of phenotype values (just one phenotype, not a matrix of them)}

\item{kinship}{Optional kinship matrix, or a list of kinship matrices (one
per chromosome), in order to use the LOCO (leave one chromosome
out) method.}

\item{addcovar}{An optional matrix of additive covariates.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{weights}{An optional vector of positive weights for the
individuals. As with the other inputs, it must have \code{names}
for individual identifiers. Ignored if \code{kinship} is provided.}

\item{contrasts}{An optional matrix of genotype contrasts, size
genotypes x genotypes. For an intercross, you might use
\code{cbind(c(1,0,0), c(-0.5, 0, 0.5), c(-0.5, 1, 0.5))} to get
mean, additive effect, and dominance effect. The default is the
identity matrix.}

\item{se}{If TRUE, also calculate the standard errors.}

\item{hsq}{(Optional) residual heritability; used only if
\code{kinship} provided.}

\item{reml}{If \code{kinship} provided: if \code{reml=TRUE}, use
REML; otherwise maximum likelihood.}

\item{tol}{Tolerance value for
linear regression by QR decomposition (in determining whether
columns are linearly dependent on others and should be omitted)}
}
\value{
A list containing the following
\itemize{
\item \code{coef} - A matrix of estimated regression coefficients, of dimension
    positions x number of effects. The number of effects is
    \code{n_genotypes + n_addcovar + (n_genotypes-1)*n_intcovar}.
\item \code{map} - A list containing the map positions at which the
    calculations were performed, taken from the input \code{genoprobs}.
\item \code{SE} - Present if \code{se=TRUE}: a matrix of estimated
    standard errors, of same dimension as \code{coef}.
\item \code{contrasts} - The input matrix of genotype coefficient
    contrasts that were used.
\item \code{addcovar} - Names of additive covariates that were used.
\item \code{intcovar} - Names of interactive covariates that were used.
\item \code{sample_size} - Vector of sample sizes used for each
    phenotype
\item \code{weights} - Logical value indicating whether weights
    were used (only included if \code{kinship} was not).
}
}
\description{
Calculate QTL effects in scan along one chromosome with a
single-QTL model using Haley-Knott regression, with possible
allowance for covariates.
}
\details{
For each of the inputs, the row names are used as
individual identifiers, to align individuals.

If \code{kinship} is absent, Haley-Knott regression is performed.
If \code{kinship} is provided, a linear mixed model is used, with a
polygenic effect estimated under the null hypothesis of no (major)
QTL, and then taken as fixed as known in the genome scan.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno[,1]
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)

# calculate coefficients for chromosome 7
coef <- scan1coef(probs[,7], pheno, addcovar=covar)

# leave-one-chromosome-out kinship matrix for chr 7
kinship7 <- calc_kinship(probs, "loco")[[7]]

# calculate coefficients for chromosome 7, adjusting for residual polygenic effect
coef <- scan1coef(probs[,7], pheno, kinship7, addcovar=covar)


}
\references{
Haley CS, Knott SA (1992) A simple
regression method for mapping quantitative trait loci in line
crosses using flanking markers.  Heredity 69:315--324.
}

