% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1perm.R
\name{scan1perm}
\alias{scan1perm}
\title{Permutation test for genome scan with a single-QTL model}
\usage{
scan1perm(genoprobs, pheno, kinship = NULL, addcovar = NULL,
  Xcovar = NULL, intcovar = NULL, weights = NULL, reml = TRUE,
  n_perm = 1, perm_Xsp = FALSE, perm_strata = NULL, chr_lengths = NULL,
  cores = 1, ...)
}
\arguments{
\item{genoprobs}{Genotype probabilities as calculated by
\code{\link[qtl2geno]{calc_genoprob}}.}

\item{pheno}{A matrix of phenotypes, individuals x phenotypes.}

\item{kinship}{Optional kinship matrix, or a list of kinship matrices (one
per chromosome), in order to use the LOCO (leave one chromosome
out) method.}

\item{addcovar}{An optional matrix of additive covariates.}

\item{Xcovar}{An optional matrix with additional additive covariates used for
null hypothesis when scanning the X chromosome.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{weights}{An optional vector of positive weights for the
individuals. As with the other inputs, it must have \code{names}
for individual identifiers. Ignored if \code{kinship} is provided.}

\item{reml}{If \code{kinship} provided: if \code{reml=TRUE}, use
REML; otherwise maximum likelihood.}

\item{n_perm}{Number of permutation replicates.}

\item{perm_Xsp}{If TRUE, do separate permutations for the autosomes
and the X chromosome.}

\item{perm_strata}{Vector of strata, for a stratified permutation
test. Should be named in the same way as the rows of
\code{pheno}. The unique values define the strata.}

\item{chr_lengths}{Lengths of the chromosomes; needed only if
\code{perm_Xsp=TRUE}. See \code{\link{chr_lengths}}.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}

\item{...}{Additional control parameters; see Details.}
}
\value{
If \code{perm_Xsp=FALSE}, the result is matrix of
genome-wide maximum LOD scores, permutation replicates x
phenotypes. If \code{perm_Xsp=TRUE}, the result is a list of
two matrices, one for the autosomes and one for the X
chromosome.
}
\description{
Permutation test for a enome scan with a single-QTL model by
Haley-Knott regression or a linear mixed model, with possible
allowance for covariates.
}
\details{
If \code{kinship} is not provided, so that analysis proceeds by
Haley-Knott regression, we permute the rows of the phenotype data;
the same permutations are also applied to the rows of the
covariates (\code{addcovar}, \code{Xcovar}, and \code{intcovar})
are permuted.

If \code{kinship} is provided, we instead permute the rows of the
genotype data and fit an LMM with the same residual heritability
(estimated under the null hypothesis of no QTL).

If \code{Xcovar} is provided and \code{perm_strata=NULL}, we do a
stratified permutation test with the strata defined by the rows of
\code{Xcovar}. If a simple permutation test is desired, provide
\code{perm_strata} that is a vector containing a single repeated
value.
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# insert pseudomarkers into map
map <- insert_pseudomarkers(iron$gmap, step=1)

# calculate genotype probabilities
probs <- calc_genoprob(iron, map, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# strata for permutations
perm_strata <- mat2strata(Xcovar)

# permutations with genome scan
\dontrun{
operm <- scan1(probs, pheno, addcovar=covar, Xcovar=Xcovar,
               n_perm=1000, perm_Xsp=TRUE, perm_strata=perm_strata,
               chr_lengths=chr_lengths(iron$gmap))}
\dontshow{operm <- scan1(probs, pheno, addcovar=covar, Xcovar=Xcovar, n_perm=3, perm_strata=perm_strata)}

# leave-one-chromosome-out kinship matrices
kinship <- calc_kinship(probs, "loco")

# genome scan with a linear mixed model
\dontrun{
operm_lmm <- scan1(probs, pheno, kinship, covar, Xcovar, n_perm=1000,
                   perm_Xsp=TRUE, perm_strata=perm_strata)}
\dontshow{operm_lmm <- scan1(probs, pheno, kinship, covar, Xcovar, n_perm=3, perm_strata=perm_strata)}

}
\references{
Churchill GA, Doerge RW (1994) Empirical threshold
values for quantitative trait mapping. Genetics 138:963--971.

Manichaikul A, Palmer AA, Sen S, Broman KW (2007) Significance
thresholds for quantitative trait locus mapping under selective
genotyping. Genetics 177:1963--1966.

Haley CS, Knott SA (1992) A simple regression method for mapping
quantitative trait loci in line crosses using flanking markers.
Heredity 69:315--324.

Kang HM, Zaitlen NA, Wade CM, Kirby A, Heckerman D, Daly MJ, Eskin
E (2008) Efficient control of population structure in model
organism association mapping. Genetics 178:1709--1723.
}
\seealso{
\code{\link{scan1}}, \code{\link{chr_lengths}}, \code{\link{mat2strata}}
}
