% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_geno.R
\name{sim_geno}
\alias{sim_geno}
\title{Simulate genotypes given observed marker data}
\usage{
sim_geno(cross, n_draws = 1, step = 0, off_end = 0,
  stepwidth = c("fixed", "max"), pseudomarker_map, error_prob = 0.0001,
  map_function = c("haldane", "kosambi", "c-f", "morgan"), quiet = TRUE,
  cores = 1)
}
\arguments{
\item{cross}{Object of class \code{"cross2"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}.}

\item{n_draws}{Number of simulations to perform.}

\item{step}{Distance between pseudomarkers and markers; if
\code{step=0} no pseudomarkers are inserted.}

\item{off_end}{Distance beyond terminal markers in which to insert
pseudomarkers.}

\item{stepwidth}{Indicates whether to use a fixed grid
(\code{stepwidth="fixed"}) or to use the maximal distance between
pseudomarkers to ensure that no two adjacent markers/pseudomarkers
are more than \code{step} apart.}

\item{pseudomarker_map}{A map of pseudomarker locations; if provided the
\code{step}, \code{off_end}, and \code{stepwidth} arguments are
ignored.}

\item{error_prob}{Assumed genotyping error probability}

\item{map_function}{Character string indicating the map function to
use to convert genetic distances to recombination fractions.}

\item{quiet}{If \code{FALSE}, print progress messages.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel]{detectCores}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel]{makeCluster}}.}
}
\value{
A list of containing the following:
\itemize{
\item \code{draws} - List of three-dimensional arrays of imputed genotypes,
individuals x positions x draws.
\item \code{map} - The genetic map as a list of vectors of marker positions.
\item \code{grid} - A list of logical vectors, indicating which
    positions correspond to a grid of markers/pseudomarkers. (may be
    absent)
\item \code{indID} - Vector of character strings with individual IDs
\item \code{chrID} - Vector of character strings with chromosome IDs
\item \code{crosstype} - The cross type of the input \code{cross}.
\item \code{is_x_chr} - Logical vector indicating whether chromosomes
    are to be treated as the X chromosome or not, from input \code{cross}.
\item \code{sex} - Vector of sexes of the individuals, from input
    \code{cross}.
\item \code{cross_info} - Matrix of cross information for the
    individuals, from input \code{cross}.
\item \code{alleles} - Vector of allele codes, from input
    \code{cross}.
\item \code{alleleprob} - Logical value (\code{FALSE}) that
    indicates whether the probabilities are compressed to allele
    probabilities, as from \code{\link{genoprob_to_alleleprob}}.
\item \code{step} - the value of the \code{step} argument.
\item \code{off_end} - the value of the \code{off_end} argument.
\item \code{stepwidth} - the value of the \code{stepwidth} argument.
}
}
\description{
Uses a hidden Markov model to simulate from the joint distribution
Pr(g | O) where g is the underlying sequence of true genotypes and
O is the observed multipoint marker data, with possible allowance
for genotyping errors.
}
\details{
After performing the backward equations, we draw from
 \eqn{Pr(g_1 = v | O)}{Pr(g[1] = v | O)} and then \eqn{Pr(g_{k+1} = v |
   O, g_k = u)}{Pr(g[k+1] = v | O, g[k] = u)}.
}
\examples{
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
draws <- sim_geno(grav2, n_draws=4, step=1, error_prob=0.002)
}
\keyword{utilities}

