% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_snps.R
\name{top_snps}
\alias{top_snps}
\title{Create table of top snp associations}
\usage{
top_snps(scan1_output, drop = 1.5, show_all_snps = TRUE)
}
\arguments{
\item{scan1_output}{Output of \code{\link[qtl2scan]{scan1}}.
Should contain a component \code{"snpinfo"}, as when
\code{\link[qtl2scan]{scan1}} is run with SNP probabilities
produced by \code{\link[qtl2scan]{genoprob_to_snpprob}}.}

\item{drop}{Show all SNPs with LOD score within this amount of the
maximum SNP association.}

\item{show_all_snps}{If TRUE, expand to show all SNPs.}
}
\description{
Create a table of the top snp associations
}
\examples{
\dontrun{
# load example DO data from web
library(qtl2geno)
file <- paste0("https://raw.githubusercontent.com/rqtl/",
               "qtl2data/master/DOex/DOex.zip")
DOex <- read_cross2(file)

# subset to chr 2
DOex <- DOex[,"2"]

# calculate genotype probabilities and convert to allele probabilities
pr <- calc_genoprob(DOex, error_prob=0.002)
apr <- genoprob_to_alleleprob(pr)

# download snp info from web
tmpfile <- tempfile()
file <- paste0("https://raw.githubusercontent.com/rqtl/",
               "qtl2data/master/DOex/c2_snpinfo.rds")
download.file(file, tmpfile, quiet=TRUE)
snpinfo <- readRDS(tmpfile)
unlink(tmpfile)

# calculate strain distribution patterns
snpinfo$sdp <- calc_sdp(snpinfo[,-(1:4)])

# switch map in allele probabilities to Mbp
apr$map <- DOex$pmap

# convert to snp probabilities
snppr <- genoprob_to_snpprob(apr, snpinfo)

# perform SNP association analysis (here, ignoring residual kinship)
out_snps <- scan1(snppr, DOex$pheno)

# table with top SNPs
top_snps(out_snps)

# top SNPs among the distinct subset at which calculations were performed
top_snps(out_snps, show_all_snps=FALSE)

# top SNPs within 1.0 LOD
top_snps(out_snps, 1.0)
}
}
\seealso{
\code{\link{genoprob_to_snpprob}}, \code{\link[qtl2plot]{plot_snpasso}}
}

