% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zip_datafiles}
\alias{zip_datafiles}
\title{Zip a set of data files}
\usage{
zip_datafiles(yaml_file, zip_file)
}
\arguments{
\item{yaml_file}{Character string with path to the yaml file
containing all of the control information.}

\item{zip_file}{Name of zip file to use. If omitted, we use the
stem of \code{yaml_file} but with a \code{.zip} extension.}
}
\value{
Character string with the file name of the zip file that
was created.
}
\description{
Zip a set of data files (in format read by \code{\link{read_cross2}}).
}
\details{
The input \code{yaml_file} is the control file (in
\href{http://www.yaml.org}{YAML} format for data to in the format
to be read by \code{\link{read_cross2}}.  (See the
\href{http://kbroman.org/qtl2/pages/sampledata.html}{sample data files} and the
\href{http://kbroman.org/qtl2/assets/vignettes/input_files.html}{vignette describing the input file format}.)

The \code{\link[utils]{zip}} function is used to do the zipping.
}
\examples{
\dontrun{
yaml_file <- "~/grav2_data/grav2.yaml"
zip_datafiles(yaml_file, "grav2.zip")
}
}
\seealso{
\code{\link{read_cross2}}, sample data files at \url{http://kbroman.org/qtl2/pages/sampledata.html}
}
\keyword{IO}

